Moved factory to first level
